requirements.txt
numpy==2.1.3
pinecone==5.3.1
python-dotenv==1.0.1
sentence_transformers==3.2.1
streamlit==1.39.0
pandas==2.2.3